// app/src/generated/models/RecommendationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface LabelValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface RecStatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AnalystRecValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface RecommendationsBase {
  ID?: number;
  Title: string;
  SOPId?: number;
  StepRef?: string;
  "Label#Id"?: number;
  Reason?: string;
  SuggestedTool?: string;
  ValueScore?: number;
  EffortScore?: number;
  Priority?: number;
  "RecStatus#Id"?: number;
  RejectionReason?: string;
  StarterFlow?: string;
  "AnalystRec#Id"?: number;
  AnalystNote?: string;
  ProposedChange?: string;
  OwnerNote?: string;
  BatchNo?: number;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface RecommendationsRead extends RecommendationsBase {
  Label?: LabelValue;
  RecStatus?: RecStatusValue;
  AnalystRec?: AnalystRecValue;
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface RecommendationsWrite extends RecommendationsBase {
  Label?: string;
  RecStatus?: string;
  AnalystRec?: string;
  Author?: AuthorValue;
  Editor?: EditorValue;
}